- separate out player card from player list component (#65)
